core/skills/deepcamera/skills/channels/matrix/requirements.txt
# Matrix Channel
matrix-nio>=0.24.0
aiohttp>=3.9.0
